Show temp credentials if DEBUG flag is on
I thought it worked, but there was dead code (views.py).

ALLOWED_HOSTS must be set if DEBUG = False. See:
https://docs.djangoproject.com/en/5.0/ref/settings/#allowed-hosts

"Allow the debug() context processor to add some variables to the
template context." See:
https://docs.djangoproject.com/en/5.0/ref/settings/#internal-ips

Extra context for the template can be passed in the urlpatterns
definition. See:
https://docs.djangoproject.com/en/5.0/topics/http/urls/#views-extra-options
as well as
https://stackoverflow.com/a/30654179